% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stack_api.R
\name{add_stack_api}
\alias{add_stack_api}
\title{Add a StackOverflow API Key}
\usage{
add_stack_api()
}
\value{
Prints instructions to the command line.
}
\description{
Checks if a StackOverflow API is present in the environment and if not
gives instructions on how to create one.
}
\examples{

add_stack_api()
}
