% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_twitter_api.R
\name{add_twitter_api}
\alias{add_twitter_api}
\title{Add a Twitter API Key}
\usage{
add_twitter_api()
}
\value{
Prints instructions to the command line.
}
\description{
Checks if a Twitter token is present at the supplied link and if not
gives instructions on how to create one.
}
\examples{

add_twitter_api()
}
