% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stack_questions.R
\name{get_stack_questions}
\alias{get_stack_questions}
\title{Get StackOverflow Questions For A List of Tags}
\usage{
get_stack_questions(extracted_tags = NULL, excluded_tags = NULL,
  time_window = 60, add_process_fn = NULL)
}
\arguments{
\item{extracted_tags}{A character vector of tags to extract questions for.}

\item{excluded_tags}{A character vector of tags to exclude questions from the
output.}

\item{time_window}{Numeric, the timewindow to extract questions for (in minutes.
Defaults to 60 minutes. Note that a maximum of 100 questions are extracted by
the StackOverflow api, if you expect your tags to have a high volume of questions
reduce this. If using this function as part of a sheduled bot makes sure this time
frame corresponds the frequency with which your bot is running.}

\item{add_process_fn}{A function. If additional munging or data extraction is required add
to a custom function and pass here. This function must except a dataframe as its first arguement
and return a dataframe. See the examples for details.}
}
\value{
A list containing the title, creation date, link, and tags for StackOverflow
questions related to your selected tags.
}
\description{
Extract StackOverflow questions for a given set of tags, exluding questions
related to an optional list of tags. This code is based on the \href{https://github.com/wjakethompson/tidyverse-tweets/blob/master/tidyversetweets.R}{tidyverse tweets bot}
by \href{https://www.wjakethompson.com/}{Jake Thompson}.
}
\examples{

## Basic call
get_stack_questions(extracted_tags = "ggplot2",
                    excluded_tags = "plotly",
                    time_window = 120)

## Add additional processing step
library(dplyr)
add_process_fn <- function(df) {
      df \%>\%
      mutate(id = 1:n())
}

get_stack_questions(extracted_tags = "ggplot2",
                    excluded_tags = "plotly",
                    time_window = 120,
                    add_process_fn = add_process_fn)
}
