% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_stack_tweets.R
\name{post_stack_tweets}
\alias{post_stack_tweets}
\title{Post A Tweet for Each Question}
\usage{
post_stack_tweets(questions = NULL, hashtags = "rstats",
  twitter_token_path = "~/twitter_token.rds", post = TRUE)
}
\arguments{
\item{questions}{A list of questions to tweet as produced by \code{\link[StackTweetBot]{get_stack_questions}}.}

\item{hashtags}{A character vector of hashtags to attach to the tweet.}

\item{twitter_token_path}{A character string indicting the path your twitter token.
See \code{\link[StackTweetBot]{add_twitter_api}} for details.}

\item{post}{Logical, defaults to \code{TRUE}. Should the formatted tweets be published on
twitter, if \code{FALSE} then the tweets are instead returned as a list.}
}
\value{
Either nothing is returned or a formated list of tweets.
}
\description{
Format and post StackOverflow questions as tweets, optionally returning them
as a formatted list. This code is based on the \href{https://github.com/wjakethompson/tidyverse-tweets/blob/master/tidyversetweets.R}{tidyverse tweets bot}
by \href{https://www.wjakethompson.com/}{Jake Thompson}.
}
\examples{

ggplot2_qs <-  get_stack_questions(extracted_tags = "ggplot2",
                                  excluded_tags = "plotly",
                                  time_window = 120)

post_stack_tweets(ggplot2_qs, hashtags = "rstats", post = FALSE)

}
